move misplaced TimedeltaIndex.delete tests
